Breaking changes on member list interfaces
This commit breaks the old member list API in order to change several
things. These changes are done to improve the overall usability and
flexibility of their interfaces.

This commit introduces the method LoadMore into the MemberListSection
interface. The method was added to allow sections to be lazy loaded
conditionally. This allows the server to lazy load the member list over
the network.

In the future, there might be further modifications to allow clients to
invalidate or mark sections as stale, which would allow the server to
free up the list.

This commit also changed the old Name method of MemberListSection to use
the unified Namer interface. It also moved from using section's contents
as identifiers to having a dedicated ID method using the unified
Identifier interface.